Turn server into executor
