+++ lib/ruby_lsp/internal.rb
@@ -20,13 +20,14 @@ require "prism/visitor"
 require "language_server-protocol"
 
 require "ruby-lsp"
+require "ruby_lsp/base_server"
 require "ruby_indexer/ruby_indexer"
 require "core_ext/uri"
 require "ruby_lsp/utils"
 require "ruby_lsp/parameter_scope"
 require "ruby_lsp/server"
-require "ruby_lsp/executor"
 require "ruby_lsp/requests"
+require "ruby_lsp/dependency_detector"
 require "ruby_lsp/response_builders"
 require "ruby_lsp/document"
 require "ruby_lsp/ruby_document"
